requirements.txt
TTS
pydub
tqdm
soundfile
spacy
python-dotenv
noisereduce 
librosa 
scipy 
numpy
pyloudnorm